requirements.txt
kiteconnect==4.2.0
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
matplotlib==3.9.1
